fix(ci): the symlink test loses its build tag, because the shard gate is right
Two integration shards went red on the previous push, and on my own file:

    FAIL: backend/internal/compose/setuptoken_unix_test.go carries build
          constraint '!windows' — not one the static shard discovery knows

The tag was chosen to avoid a runtime skip, on the grounds that a security check
which skips reads exactly like one that passes. That reasoning still holds; the
mechanism was the wrong one, and test-integration-parallel.sh explains why in the
comment above the allowlist it refused me from: a satisfied built-in tag compiles
under `go test` yet lands in no slice, and the ran==assigned teeth cannot see a
test that discovery itself never counted. On the Linux shard runners `!windows`
IS satisfied. So the file would have compiled and been invisible.

Teaching the gate that tag would have been the easy fix and an unsafe one. The
allowlist is a set of constraints this lane's build DECIDES — `integration` it
always sets, `e2e_llm`/`livesmoke`/`voicelive` it never does. `!windows` is in
neither category: it is satisfied, so a future integration test carrying it would
be silently dropped from every shard while passing this scan. The gate is not
being conservative for its own sake, and it should keep refusing.

So the test moves into the untagged setuptoken_test.go and skips on Windows at
runtime, for the FIXTURE rather than the property: planting a symlink there needs
SeCreateSymbolicLinkPrivilege or Developer Mode. Every lane that runs this suite
runs on Linux, so it never skips in CI — which is the concern the tag was
reaching for, satisfied by where the lanes actually run rather than by a
constraint expression. The Windows behaviour it cannot cover is the dangling
final symlink, already named in setuptokenflags_windows.go and tracked in #1579.

Verified by running the gate rather than reasoning about it: shard 5/12 now
discovers 231 of 2769 tests across 35 packages and gets as far as needing a
database, which is where a run without `make db-up` is supposed to stop. The test
still executes in the unit lane — both subtests RUN and PASS, not skip — and
still catches the mutant that replaces O_EXCL with O_TRUNC.

The first version of that verification script reported success having executed
nothing: it wrapped the run in `timeout`, which macOS does not ship, so the
command failed instantly and the grep found no failure in an empty string. Third
time this session that a check needed checking.

Also staged the deletion rather than only removing the file: check-gofmt.sh reads
git ls-files, so a tracked path with nothing behind it fails the gate on lstat.
